# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LoadBoardPro API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /loads:
    get:
      operationId: listLoads
      tags: [loads]
      summary: List loads with filtering and pagination
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: driverId
          in: query
          schema:
            type: string
        - name: dispatcherId
          in: query
          schema:
            type: string
        - name: brokerId
          in: query
          schema:
            type: string
        - name: weekStart
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Loads list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoadsPage"
    post:
      operationId: createLoad
      tags: [loads]
      summary: Create a new load
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Load"
  /loads/{id}:
    get:
      operationId: getLoad
      tags: [loads]
      summary: Get load details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Load
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Load"
    patch:
      operationId: updateLoad
      tags: [loads]
      summary: Update a load
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoadUpdate"
      responses:
        "200":
          description: Updated load
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Load"
    delete:
      operationId: deleteLoad
      tags: [loads]
      summary: Soft delete a load (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── ANALYTICS / KPI ──────────────────────────────────────────────────────
...
  /accounting/summary:
    get:
      operationId: getAccountingSummary
      tags: [loads]
      summary: Get accounting summary metrics
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Accounting summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountingSummary"

  # ─── NOTIFICATIONS ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, role]
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, dispatcher, accounting, driver]
        isActive:
          type: boolean
        createdAt:
          type: string
...
    Driver:
      type: object
      required: [id, fullName, driverType, isActive]
      properties:
        id:
          type: string
        fullName:
          type: string
        driverType:
          type: string
          enum: [OO, CD, Lease]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        truckNumber:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
...
    Broker:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
        mcNumber:
          type: ["string", "null"]
        contact:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Load:
      type: object
      required: [id, loadNumber, puDate, delDate, originCity, originState, destCity, destState, mileage, rate, status, weekStart]
      properties:
        id:
          type: string
        loadNumber:
          type: string
        driverId:
          type: ["string", "null"]
        driver:
          $ref: "#/components/schemas/Driver"
        dispatcherId:
          type: ["string", "null"]
        dispatcher:
          $ref: "#/components/schemas/User"
        brokerId:
          type: ["string", "null"]
        broker:
          $ref: "#/components/schemas/Broker"
        puDate:
          type: string
        delDate:
          type: string
        originCity:
          type: string
        originState:
          type: string
        destCity:
          type: string
        destState:
          type: string
        mileage:
          type: number
        rate:
          type: number
        rpm:
          type: ["number", "null"]
        status:
          type: string
          enum: [Booked, InQM, Delivered, Canceled, Completed, NeedRevRC, Issue, Checked, Invoiced, Reinvoiced, BrokerPaid]
        reimbursement:
          type: number
        dispatchNotes:
          type: ["string", "null"]
        invoicedAmount:
          type: ["number", "null"]
        brokerPaid:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        weekStart:
          type: string
        irDiff:
          type: ["number", "null"]
        biDiff:
          type: ["number", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    LoadInput:
      type: object
      required: [loadNumber, puDate, delDate, originCity, originState, destCity, destState, mileage, rate, weekStart]
      properties:
        loadNumber:
          type: string
        driverId:
          type: ["string", "null"]
        dispatcherId:
          type: ["string", "null"]
        brokerId:
          type: ["string", "null"]
        puDate:
          type: string
        delDate:
          type: string
        originCity:
          type: string
        originState:
          type: string
        destCity:
          type: string
        destState:
          type: string
        mileage:
          type: number
        rate:
          type: number
        status:
          type: string
          enum: [Booked, InQM, Delivered, Canceled, Completed, NeedRevRC, Issue, Checked, Invoiced, Reinvoiced, BrokerPaid]
        reimbursement:
          type: number
        dispatchNotes:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        weekStart:
          type: string
    LoadUpdate:
      type: object
      properties:
        loadNumber:
          type: string
        driverId:
          type: ["string", "null"]
        dispatcherId:
          type: ["string", "null"]
        brokerId:
          type: ["string", "null"]
        puDate:
          type: string
        delDate:
          type: string
        originCity:
          type: string
        originState:
          type: string
        destCity:
          type: string
        destState:
          type: string
        mileage:
          type: number
        rate:
          type: number
        status:
          type: string
          enum: [Booked, InQM, Delivered, Canceled, Completed, NeedRevRC, Issue, Checked, Invoiced, Reinvoiced, BrokerPaid]
        reimbursement:
          type: number
        dispatchNotes:
          type: ["string", "null"]
        invoicedAmount:
          type: ["number", "null"]
        brokerPaid:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
    LoadsPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Load"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── ANALYTICS ───────────────────────────────────────────────────────────
...
    AccountingSummary:
      type: object
      properties:
        totalInvoiced:
          type: number
        brokerPaidThisWeek:
          type: number
        outstanding:
          type: number
        diffIssues:
          type: integer

    # ── NOTIFICATIONS ────────────────────────────────────────────────────────
...